# docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/v1/device/admin/update": {
            "put": {
                "tags": [
                    "device-controller"
                ],
                "operationId": "updateData",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/device/admin/delete/{id}": {
            "put": {
                "tags": [
                    "device-controller"
                ],
                "operationId": "updateData_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/windspeed/create": {
            "post": {
                "tags": [
                    "wind-speed-controller"
                ],
                "operationId": "createData",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/admin/create": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "createData_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/solarradiation/create": {
            "post": {
                "tags": [
                    "solar-radiation-controller"
                ],
                "operationId": "createData_2",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/soilph/create": {
            "post": {
                "tags": [
                    "soil-ph-controller"
                ],
                "operationId": "createData_3",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/soilmoisture/create": {
            "post": {
                "tags": [
                    "soil-moisture-controller"
                ],
                "operationId": "createData_4",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rainfall/create": {
            "post": {
                "tags": [
                    "rain-fall-controller"
                ],
                "operationId": "createData_5",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/humidity/create": {
            "post": {
                "tags": [
                    "humidity-controller"
                ],
                "operationId": "createData_6",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/device/admin/create": {
            "post": {
                "tags": [
                    "device-controller"
                ],
                "operationId": "createData_7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/airtemperature/create": {
            "post": {
                "tags": [
                    "air-temperature-controller"
                ],
                "operationId": "createData_8",
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/windspeed/lastdata": {
            "get": {
                "tags": [
                    "wind-speed-controller"
                ],
                "operationId": "readAllData",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/windspeed/alldata": {
            "get": {
                "tags": [
                    "wind-speed-controller"
                ],
                "operationId": "readData",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/solarradiation/lastdata": {
            "get": {
                "tags": [
                    "solar-radiation-controller"
                ],
                "operationId": "readAllData_1",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/solarradiation/alldata": {
            "get": {
                "tags": [
                    "solar-radiation-controller"
                ],
                "operationId": "readData_1",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/soilph/lastdata": {
            "get": {
                "tags": [
                    "soil-ph-controller"
                ],
                "operationId": "readAllData_2",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/soilph/alldata": {
            "get": {
                "tags": [
                    "soil-ph-controller"
                ],
                "operationId": "readData_2",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/soilmoisture/lastdata": {
            "get": {
                "tags": [
                    "soil-moisture-controller"
                ],
                "operationId": "readAllData_3",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/soilmoisture/alldata": {
            "get": {
                "tags": [
                    "soil-moisture-controller"
                ],
                "operationId": "readData_3",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/rainfall/lastdata": {
            "get": {
                "tags": [
                    "rain-fall-controller"
                ],
                "operationId": "readAllData_4",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/rainfall/alldata": {
            "get": {
                "tags": [
                    "rain-fall-controller"
                ],
                "operationId": "readData_4",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DeviceRequest": {
                "required": [
                    "latitude",
                    "location",
                    "longitude",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "string"
                    }
                }
            },
            "WebResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "object"
                    },
                    "data": {
                        "type": "object"
                    },
                    "paging": {
                        "type": "object"
                    }
                }
            },
            "DataRequest": {
                "required": [
                    "deviceId",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "deviceId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "value": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "CreateUserRequest": {
                "required": [
                    "biodata",
                    "users"
                ],
                "type": "object",
                "properties": {
                    "biodata": {
                        "$ref": "#/components/schemas/BiodataRequest"
                    },
                    "users": {
                        "$ref": "#/components/schemas/UserRequest"
                    }
                }
            },
...
        }
    }
}